backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
groq
python-multipart
pydantic
python-dotenv
